Got the Magnetometer to work. It only calls up the setup function once and now works. Created a FindMagOffset executable which when running will calculate the necessary X,Y value offset (can be changed in the magnetometer.cc file). Also created a testMagnetometer executable for 4wd.

+++ 2012/software/motors/osmc/arduino/pinDefs.hpp
@@ -27,9 +27,9 @@ const int joystickEnable = 8;
 
 const int LIGHT_PIN_LB = 13;
 
-const int magCLK = 8;
-const int magEN = 9;
-const int magDIO = 10;
+const byte magCLK = 8;
+const byte magEN = 9;
+const byte magDIO = 10;
